Put service object to GoogleMap namespace
