+++ test/integration/google_maps_data_builder_service_object_test.rb
@@ -3,7 +3,7 @@ require "test_helper"
 class GoogleMapsDataBuilderServiceObjectTest < ActionDispatch::IntegrationTest
   test "should return array of hashes with latitude, longitude, dog_name, and breed" do
     adoptions = Adoption.includes(:dog, adopter_account: { adopter_profile: :location })
-    result = GoogleMapsDataBuilder.new(adoptions).data
+    result = GoogleMap::GoogleMapsDataBuilder.new(adoptions).data
 
     assert_instance_of Array, result
     assert_instance_of Hash, result.first
